refactor: PenaltyKickReferee 클래스에 있지 말아야할 메서드 Controller로 이동
